Migrated list-role/course/quarter/tags to one service function.
