+++ src/commands/list-courses/list-courses.command.ts
@@ -5,6 +5,7 @@ import {
   ICommand,
 } from "@/commands/command.interface";
 import { getGuildConfig } from "@/config/guilds.config";
+import { displayList } from "@/services/display-list.service";
 import {
   CommandInteraction,
   Message,
@@ -52,104 +53,7 @@ export class ListCoursesCommand implements ICommand {
       );
     }
 
-    displayList(i, filteredCourses, searchTerm);
+    displayList(i, filteredCourses, "Courses", searchTerm);
   }
 }
 
-const displayList = async (
-  i: CommandInteraction,
-  list: string[],
-  searchTerm?: string
-) => {
-  let pageNumber = 0;
-
-  const buttons = [
-    new MessageActionRow().addComponents([
-      new MessageButton()
-        .setCustomId("list-back")
-        .setStyle("PRIMARY")
-        .setEmoji("⬅️")
-        .setDisabled(true),
-      new MessageButton()
-        .setCustomId("list-forward")
-        .setStyle("PRIMARY")
-        .setEmoji("➡️"),
-    ]),
-  ];
-  const initialEmbed = getCourseListEmbedPage(list, pageNumber, searchTerm);
-
-  if (list.length <= 24) {
-    i.followUp({ embeds: initialEmbed });
-    return;
-  }
-
-  if (initialEmbed[0].fields.length !== 24) {
-    buttons[0].components[1].setDisabled(true);
-  }
-  const sentReplyMessage = (await i.followUp({
-    embeds: initialEmbed,
-    components: buttons,
-  })) as Message;
-
-  const collector = sentReplyMessage.createMessageComponentCollector({
-    componentType: "BUTTON",
-    time: 300000,
-  });
-  collector.on("collect", async (b) => {
-    if (b.user.id !== i.user.id) {
-      await b.reply({
-        content: "You're not allowed to do that. Make your own search.",
-        ephemeral: true,
-      });
-      return;
-    }
-    await b.reply({ content: "​" });
-    await b.deleteReply();
-    if (b.customId === "list-back") {
-      buttons[0].components[1].setDisabled(false);
-      if (pageNumber - 1 === 0) {
-        buttons[0].components[0].setDisabled(true);
-      }
-      i.editReply({
-        embeds: getCourseListEmbedPage(list, --pageNumber, searchTerm),
-        components: buttons,
-      });
-    }
-    if (b.customId === "list-forward") {
-      const embed = getCourseListEmbedPage(list, ++pageNumber, searchTerm);
-      buttons[0].components[0].setDisabled(false);
-      if (embed[0].fields.length !== 24) {
-        buttons[0].components[1].setDisabled(true);
-      }
-      i.editReply({
-        embeds: embed,
-        components: buttons,
-      });
-    }
-  });
-};
-
-function getCourseListEmbedPage(
-  courseList: string[],
-  pageNumber: number,
-  searchTerm?: string
-): MessageEmbed[] {
-  let filteredCourses = courseList.slice(
-    24 * pageNumber,
-    24 * (pageNumber + 1)
-  );
-  if (filteredCourses.length === 0) {
-    filteredCourses = courseList.slice(24 * pageNumber);
-  }
-
-  let currentEmbed = new MessageEmbed();
-  currentEmbed
-    .setTitle("Courses " + (searchTerm ? `with \`${searchTerm}\`` : ""))
-    .setColor("BLUE");
-
-  for (const c of filteredCourses) {
-    currentEmbed.addField(`\`${c}\``, "​", true);
-  }
-
-  return [currentEmbed];
-}
