+++ src/commands/list-roles/list-roles.command.ts
@@ -11,6 +11,7 @@ import {
   ICommand,
 } from "@/commands/command.interface";
 import { getGuildConfig } from "@/config/guilds.config";
+import { displayList } from "@/services/display-list.service";
 
 export class ListRolesCommand implements ICommand {
   name = "list-roles";
@@ -36,7 +37,6 @@ export class ListRolesCommand implements ICommand {
     const guildConfig = getGuildConfig(i.guildId);
 
     i.deferReply();
-    let pageNumber = 0;
     const searchTerm = i.options.getString("search", false)?.toLowerCase();
     let roles = (
       guildConfig?.selfRoles.filter((r) =>
@@ -44,90 +44,6 @@ export class ListRolesCommand implements ICommand {
       ) || []
     ).map((r) => r.name);
 
-    const buttons = [
-      new MessageActionRow().addComponents([
-        new MessageButton()
-          .setCustomId("role-list-back")
-          .setStyle("PRIMARY")
-          .setEmoji("⬅️")
-          .setDisabled(true),
-        new MessageButton()
-          .setCustomId("role-list-forward")
-          .setStyle("PRIMARY")
-          .setEmoji("➡️"),
-      ]),
-    ];
-    const initialEmbed = getRoleListEmbedPage(roles, pageNumber, searchTerm);
-
-    if (roles.length <= 24) {
-      i.followUp({ embeds: initialEmbed });
-      return;
-    }
-
-    if (initialEmbed[0].fields.length !== 24) {
-      buttons[0].components[1].setDisabled(true);
-    }
-    const sentReplyMessage = (await i.followUp({
-      embeds: initialEmbed,
-      components: buttons,
-    })) as Message;
-
-    const collector = sentReplyMessage.createMessageComponentCollector({
-      componentType: "BUTTON",
-      time: 300000,
-    });
-    collector.on("collect", async (b) => {
-      if (b.user.id !== i.user.id) {
-        await b.reply({
-          content: "You're not allowed to do that. Make your own search.",
-          ephemeral: true,
-        });
-        return;
-      }
-      await b.reply({ content: "​" });
-      await b.deleteReply();
-      if (b.customId === "role-list-back") {
-        buttons[0].components[1].setDisabled(false);
-        if (pageNumber - 1 === 0) {
-          buttons[0].components[0].setDisabled(true);
-        }
-        i.editReply({
-          embeds: getRoleListEmbedPage(roles, --pageNumber, searchTerm),
-          components: buttons,
-        });
-      }
-      if (b.customId === "role-list-forward") {
-        const embed = getRoleListEmbedPage(roles, ++pageNumber, searchTerm);
-        buttons[0].components[0].setDisabled(false);
-        if (embed[0].fields.length !== 24) {
-          buttons[0].components[1].setDisabled(true);
-        }
-        i.editReply({
-          embeds: embed,
-          components: buttons,
-        });
-      }
-    });
+    displayList(i, roles, "Self roles", searchTerm);
   }
-}
-
-function getRoleListEmbedPage(
-  roleList: string[],
-  pageNumber: number,
-  searchTerm?: string
-): MessageEmbed[] {
-  let filteredRoles = roleList.slice(24 * pageNumber, 24 * (pageNumber + 1));
-  if (filteredRoles.length === 0) {
-    filteredRoles = roleList.slice(24 * pageNumber);
-  }
-
-  let currentEmbed = new MessageEmbed();
-  currentEmbed
-    .setTitle("Roles " + (searchTerm ? `with \`${searchTerm}\`` : ""))
-    .setColor("BLUE");
-  for (const r of filteredRoles) {
-    currentEmbed.addField(`\`${r}\``, "​", true);
-  }
-
-  return [currentEmbed];
 }
